umm yeahhhh... NPCs change hats a bunch
I'm just committing this now because it's finally sort of working.  I
think there might be something wrong with players transmitting their
trends to nearby NPCs but NPC->NPC seems to work.  I had to do a bunch
of stupid stuff in order to get Unity to like my code (apparently you
cannot use ANY UnityEngine methods like Random or .renderer unless you
are in one of the main functions like Update()) so it's uglier than it
was before.